ai-service/requirements.txt
flask
flask-cors
sentence-transformers
duckduckgo-search
faiss-cpu
numpy
mysql-connector-python
requests
